feat: add registration functionality and refactor video playback logic for improved concurrency
